+++ src/Unpacker/ProductUnpacker.php
@@ -14,6 +14,7 @@ use Heptacom\HeptaConnect\Dataset\Ecommerce\Tax\TaxGroupRule;
 use Heptacom\HeptaConnect\Portal\LocalShopwarePlatform\Support\DalAccess;
 use Heptacom\HeptaConnect\Portal\LocalShopwarePlatform\Support\ExistingIdentifierCache;
 use Heptacom\HeptaConnect\Portal\LocalShopwarePlatform\Support\PrimaryKeyGenerator;
+use Heptacom\HeptaConnect\Portal\LocalShopwarePlatform\Support\TranslationLocaleCache;
 use Shopware\Core\Content\Product\Aggregate\ProductVisibility\ProductVisibilityDefinition;
 use Shopware\Core\Defaults;
 use Shopware\Core\Framework\DataAbstractionLayer\Search\Criteria;
@@ -38,6 +39,8 @@ class ProductUnpacker
 
     private ProductPriceUnpacker $productPriceUnpacker;
 
+    private TranslationLocaleCache $translationLocaleCache;
+
     private ?SalesChannelCollection $salesChannels = null;
 
     public function __construct(
@@ -46,7 +49,8 @@ class ProductUnpacker
         MediaUnpacker $mediaUnpacker,
         ManufacturerUnpacker $manufacturerUnpacker,
         UnitUnpacker $unitUnpacker,
-        ProductPriceUnpacker $productPriceUnpacker
+        ProductPriceUnpacker $productPriceUnpacker,
+        TranslationLocaleCache $translationLocaleCache
     ) {
         $this->dalAccess = $dalAccess;
         $this->existingIdentifierCache = $existingIdentifierCache;
@@ -54,6 +58,7 @@ class ProductUnpacker
         $this->manufacturerUnpacker = $manufacturerUnpacker;
         $this->unitUnpacker = $unitUnpacker;
         $this->productPriceUnpacker = $productPriceUnpacker;
+        $this->translationLocaleCache = $translationLocaleCache;
     }
 
     public function unpack(Product $source): array
@@ -124,8 +129,6 @@ class ProductUnpacker
             'purchaseSteps' => (int) \max($source->getPurchaseQuantity() ?? 1, 1),
             'isCloseout' => true,
             'shippingFree' => false,
-            'name' => $source->getName()->getFallback(),
-            'description' => $source->getDescription()->getFallback(),
             'visibilities' => $visibilities,
             'taxId' => $taxId,
             ($unit === null ? 'unitId' : 'unit') => $unit,
@@ -144,6 +147,7 @@ class ProductUnpacker
                     fn (Category $category) => $this->dalAccess->idExists('category', $category->getPrimaryKey())
                 )
             ),
+            'translations' => $this->unpackTranslations($source),
         ];
     }
 
@@ -267,5 +271,25 @@ class ProductUnpacker
         ], $unpackedMedias);
 
         return $productMedias;
+    }
+
+    protected function unpackTranslations(Product $source): array
+    {
+        $result = [];
+
+        foreach ($this->translationLocaleCache->getLocales() as $localeCode) {
+            $result[$localeCode]['name'] = null;
+            $result[$localeCode]['description'] = null;
+        }
+
+        foreach ($source->getName()->getLocaleKeys() as $localeCode) {
+            $result[$localeCode]['name'] ??= $source->getName()->getTranslation($localeCode);
+        }
+
+        foreach ($source->getDescription()->getLocaleKeys() as $localeCode) {
+            $result[$localeCode]['description'] ??= $source->getDescription()->getTranslation($localeCode);
+        }
+
+        return [];
     }
 }